move managers to their own folder and fix ssl problems
